refactor: Update AuthPage, Logout, Auth, AuthHelpers, and _requests modules
