+++ src/app/modules/_auth/Logout.tsx
@@ -1,9 +1,10 @@
-import {useEffect} from 'react'
-import {Navigate, Routes} from 'react-router-dom'
-import {useAuth} from './core/Auth'
+import { useEffect } from 'react'
+import { Navigate, Routes } from 'react-router-dom'
+import { useAuth } from './core/Auth'
 
 export function Logout() {
-  const {logout} = useAuth()
+  const { logout } = useAuth()
+
   useEffect(() => {
     logout()
     document.location.reload()
