+++ src/app/modules/_auth/core/_requests.ts
@@ -3,10 +3,10 @@ import { AuthModel, UserModel } from "./_models";
 
 const API_URL = import.meta.env.VITE_APP_API_URL;
 
-export const GET_USER_BY_ACCESSTOKEN_URL = `${API_URL}/verify_token`;
-export const LOGIN_URL = `${API_URL}/login`;
-export const REGISTER_URL = `${API_URL}/register`;
-export const REQUEST_PASSWORD_URL = `${API_URL}/forgot_password`;
+export const GET_USER_BY_ACCESSTOKEN_URL = `${API_URL}/auth/verify_token`;
+export const LOGIN_URL = `${API_URL}/auth/login`;
+export const REGISTER_URL = `${API_URL}/auth/register`;
+export const REQUEST_PASSWORD_URL = `${API_URL}/auth/forgot_password`;
 
 // Server should return AuthModel
 export function login(email: string, password: string) {
